UPDATE TO VERSION 0.4.2.15

+++ oceanbase/tests/common/serialization_test.cpp
@@ -32,7 +32,7 @@ TEST(serialization,int8)
 
   for(int i=0;i<BUF_SIZE;++i)
   {
-    ASSERT_EQ(serialization::encode_i8(buf,BUF_SIZE,pos,i),OB_SUCCESS);
+    ASSERT_EQ(serialization::encode_i8(buf,BUF_SIZE,pos,static_cast<int8_t>(i)),OB_SUCCESS);
   }
   write_buf_to_file("serialization_int8",buf,BUF_SIZE);
   ASSERT_EQ(pos,BUF_SIZE);
