+++ oceanbase/tests/sstable/ob_sstable_schemaV1.cpp
@@ -1,5 +1,5 @@
 /**
- * (C) 2010 Taobao Inc.
+ * (C) 2010-2011 Taobao Inc.
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License 
@@ -250,7 +250,7 @@ namespace oceanbase
         }
         else
         {
-          ret = middle;
+          ret = static_cast<int32_t>(middle);
           break;
         }
       }
